pyproject.toml
[project]
name = "legal-rag"
version = "0.1.0"
description = "Legal RAG System - Retrieval-Augmented Generation for Japanese Legal Documents"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "qdrant-client>=1.7.0",
    "httpx>=0.25.0",
    "tqdm>=4.66.0",
    "datasets>=2.15.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.2",
    "ragas>=0.1.0",
    "pandas>=2.0.0",
    # 日本語トークナイザー（管理者権限不要）
    "sudachipy>=0.6.0",
    "sudachidict-core>=20230927",
    "janome>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]
# MeCab（管理者権限が必要、レガシー環境用）
mecab = [
    "mecab-python3>=1.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
